move passphrase related stuff to borg.helpers.passphrase
